requirements.txt
fastapi
uvicorn
ollama
vosk
sounddevice
numpy
pydantic
psycopg2-binary
